# booking-service/src/booking/entities/read/booking-read.entity.ts
import {
  Entity,
  Column,
  PrimaryColumn,
  CreateDateColumn,
} from 'typeorm';

export enum SessionType {
  VIDEO = 'VIDEO',
  CHAT = 'CHAT',
  AUDIO = 'AUDIO',
}

export enum BookingStatus {
  PENDING = 'PENDING',
  CONFIRMED = 'CONFIRMED',
  IN_PROGRESS = 'IN_PROGRESS',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
  NO_SHOW = 'NO_SHOW',
}

export class BookingSessionRead {
  id: string;
  startTime: Date;
  endTime: Date;
  durationMinutes: number;
  sessionType: SessionType;
  sessionUrl?: string;
  recordingUrl?: string;
}

@Entity('bookings')
export class BookingRead {
  @PrimaryColumn('uuid')
  id: string;

  @Column({ type: 'uuid' })
  userId: string;

  @Column({ type: 'uuid', nullable: true })
  psychologistId: string;

  @Column({
    type: 'enum',
    enum: BookingStatus,
  })
  status: BookingStatus;

  @Column({ type: 'json', nullable: true })
  session: BookingSessionRead;

  @Column({ type: 'text', nullable: true })
  notes: string;

  @Column({ type: 'text', nullable: true })
  consultationNotes: string;

  @Column({ type: 'timestamp' })
  scheduledAt: Date;

  @CreateDateColumn({ type: 'timestamp' })
  createdAt: Date;

  @Column({ type: 'timestamp', nullable: true })
  updatedAt: Date;

  @Column({ type: 'timestamp', nullable: true })
  syncedAt: Date;
}